requirements.txt
Flask>=3.0.0,<4.0.0
requests>=2.32.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
livereload>=2.6.0,<3.0.0
